+++ CHANGELOG.md
@@ -10,6 +10,18 @@
   div{color:rgba(0,0,0,0.5);}
   ```
 
+* Removed last semicolon from declaration groups in compact output.
+
+  Before:
+  ```css
+  div{color:red;background:blue;}
+  ```
+
+  After:
+  ```css
+  div{color:red;background:blue}
+  ```
+
 ## 0.14.5
 
 * Fixed a crashed caused by parsing `:host()` without an argument and added an